+++ libs/common/src/lib/helper.ts
@@ -82,6 +82,7 @@ const USER_SETTINGS_KEYS_OF_AUTHENTICATED_USER: (keyof UserSettings)[] = [
   'isRestrictedView',
   'language',
   'locale',
+  'overviewChartMode',
   'viewMode'
 ];